+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ### Changed
 - Added limitations to methodology ([#27](https://gitlab.heigit.org/climate-action/plugins/traffic-emissions/-/issues/27))
 - Normalize legend colors logarithmically so the different values are better recognisable in the map ([#29](https://gitlab.heigit.org/climate-action/plugins/traffic-emissions/-/issues/29))
+- Make plugin usable Germany-wide by integrating raster data download from Google Earth Engine ([#24](https://gitlab.heigit.org/climate-action/plugins/traffic-emissions/-/issues/24))
 
 ## [0.1.0](https://gitlab.heigit.org/climate-action/plugins/traffic-emissions/-/releases/0.1.0) - 2024-06-25